0.8.7.7
Whole thing has been themed with scanlines and a blur for that familiar PipBoy theme. Doing this has cause rendering issues with scrollable menus like inventory etc, the touch area is still working but the items that should be hidden as you scroll are instead being drawn on top of everything else. I think this is an issue with the EffectWidget, and I've submitted a bug report to Kivy. Either way, I'm looking at ways to fix it myself if I can.